removeDuplicatedRegions: change from O((n+m)^2) to O(n*m + m^2).
The usual case is number of previous rectangles being big and number of new
rectangles being small. As the function is always called after adding more
rectangles, one can assume that the set of pre-existing rectangles has no
rectangles containing other rectangles, and thus that one only needs to
search for duplicates when one of the two rectangles is a newly added one.

This speeds up AAAAXY loading on a Moto G7 Play from 52.27 seconds to
8.06 seconds.